Moved field choice functions to new file as per review.
Rewrote code in FieldChoiceView more parametric per review. CATEGORY_MODELS_MAPPING in models. #658 #853
