+++ signbank/tools.py
@@ -2048,75 +2048,6 @@ def map_search_results_to_gloss_list(search_results):
     return (gloss_ids, Gloss.objects.filter(id__in=gloss_ids))
 
 
-def fields_to_categories():
-    choice_categories = []
-    for topic in ['main', 'phonology', 'semantics']:
-        for field in FIELDS[topic]:
-            if field == 'semField':
-                # capitalize it
-                if 'SemField' not in choice_categories:
-                    choice_categories.append('SemField')
-                continue
-            elif field in ['derivHist']:
-                if field not in choice_categories:
-                    choice_categories.append(field)
-                continue
-            # the following check will be used when querying is added, at the moment these don't appear in the phonology list
-            if field in settings.HANDSHAPE_ETYMOLOGY_FIELDS + settings.HANDEDNESS_ARTICULATION_FIELDS:
-                continue
-            if field in [f.name for f in Gloss._meta.fields]:
-                field_field = Gloss._meta.get_field(field)
-            elif field in [f.name for f in Handshape._meta.fields]:
-                field_field = Handshape._meta.get_field(field)
-            elif field in [f.name for f in Morpheme._meta.fields]:
-                field_field = Morpheme._meta.get_field(field)
-            else:
-                print('field_to_categories: field not found in Gloss, Handshape, Morpheme: ', field)
-                continue
-            if field in ['domhndsh', 'subhndsh', 'final_domhndsh', 'final_subhndsh']:
-                if 'Handshape' not in choice_categories:
-                    choice_categories.append('Handshape')
-            elif hasattr(field_field, 'field_choice_category'):
-                if field_field.field_choice_category not in choice_categories:
-                    choice_categories.append(field_field.field_choice_category)
-    return choice_categories
-
-def fields_to_fieldcategory_dict(fields=[]):
-    # allows an optional fields parameter for cases where FieldChoiceForeignKey fields from other classes are needed
-    # Normally only Gloss fields are needed
-    if not fields:
-        fields = FIELDS['main'] + FIELDS['phonology'] + FIELDS['semantics']
-    choice_categories = {}
-    for field in fields:
-        if field == 'semField':
-            choice_categories[field] = 'SemField'
-            continue
-        elif field in ['derivHist']:
-            choice_categories[field] = 'derivHist'
-            continue
-        # the following check will be used when querying is added, at the moment these don't appear in the phonology list
-        if field in settings.HANDSHAPE_ETYMOLOGY_FIELDS + settings.HANDEDNESS_ARTICULATION_FIELDS:
-            continue
-        if field in [f.name for f in Gloss._meta.fields]:
-            field_field = Gloss._meta.get_field(field)
-        elif field in [f.name for f in Handshape._meta.fields]:
-            field_field = Handshape._meta.get_field(field)
-        elif field in [f.name for f in Morpheme._meta.fields]:
-            field_field = Morpheme._meta.get_field(field)
-        elif field in [f.name for f in Definition._meta.fields]:
-            field_field = Definition._meta.get_field(field)
-        elif field in [f.name for f in OtherMedia._meta.fields]:
-            field_field = OtherMedia._meta.get_field(field)
-        elif field in [f.name for f in MorphologyDefinition._meta.fields]:
-            field_field = MorphologyDefinition._meta.get_field(field)
-        else:
-            print('field_to_categories: field not found in Gloss, Handshape, Morpheme: ', field)
-            continue
-        if field in ['domhndsh', 'subhndsh', 'final_domhndsh', 'final_subhndsh']:
-            choice_categories[field] = 'Handshape'
-        elif hasattr(field_field, 'field_choice_category'):
-            choice_categories[field] = field_field.field_choice_category
-    return choice_categories
 
 def get_interface_language_and_default_language_codes(request):
     default_language = Language.objects.get(id=get_default_language_id())
